# backend/src/test/resources/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 교원사이트 공통기능 API
  version: 1.0.0
  description: CMS-1351 test0731 1차 완료 목표 HTTP API 계약
servers:
- url: /
paths:
...
  /api/roles/{roleCode}:
    put:
      operationId: updateRole
      summary: 역할 기준 수정
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-003
      - REQ-017
      - REQ-021
      - REQ-027
      x-business-rules:
      - 존재하지 않는 role_code 수정은 거부한다
      - 역할명 변경 후에도 role_code는 변경하지 않는다
      x-side-effects:
      - role의 purpose/assignment_criteria/default_data_scope를 갱신한다
      - change_history에 역할 수정 전후 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 역할 기준 수정 요청 -> 후속 조회에서 관련 table 변경값이 ApiResponse.data에 표시된다'
      - 'auth: 세션 없는 역할 기준 수정 요청 -> HTTP 401과 ApiError.message가 반환된다'
      - 'validation: 필수 식별자 누락 -> ApiError.fields에 해당 field 오류가 포함된다'
      - 'business: 존재하지 않는 role_code 수정은 거부한다 위반 -> 대상 table의 기존 상태가 변경되지 않는다'
      - 'side-effect: 역할 기준 수정 성공 -> change_history 또는 관련 상태 row가 생성된다'
      responses:
        '200':
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 검증 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      parameters:
      - name: roleCode
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      x-roles:
      - authenticated
  /api/user-roles:
    get:
      operationId: listUserRoles
      summary: 사용자 역할 조회
      x-related-requirements:
      - REQ-004
      - REQ-027
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 검증 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - sessionCookie: []
      parameters: *id004
      x-roles:
      - authenticated
    post:
      operationId: grantUserRole
      summary: 수동 역할 부여
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-004
      - REQ-017
      - REQ-021
      - REQ-027
      x-business-rules:
      - assignment_type=MANUAL 역할만 시스템관리자가 부여할 수 있다
      - effective_end_date는 시작일보다 이전일 수 없다
      x-side-effects:
      - user_role row를 ACTIVE 상태로 생성한다
      - change_history에 승인자와 유효기간을 기록한다
      x-required-tests:
      - 'happy: 유효한 수동 역할 부여 요청 -> 후속 조회에서 관련 table 변경값이 ApiResponse.data에 표시된다'
      - 'auth: 세션 없는 수동 역할 부여 요청 -> HTTP 401과 ApiError.message가 반환된다'
      - 'validation: 필수 식별자 누락 -> ApiError.fields에 해당 field 오류가 포함된다'
      - 'business: assignment_type=MANUAL 역할만 시스템관리자가 부여할 수 있다 위반 -> 대상 table의 기존 상태가 변경되지 않는다'
      - 'side-effect: 수동 역할 부여 성공 -> change_history 또는 관련 상태 row가 생성된다'
      responses:
        '200':
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 검증 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      x-state-transitions:
      - 없음 -> ACTIVE
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      x-roles:
      - authenticated
  /api/user-roles/{assignmentId}:
    put:
      operationId: updateUserRole
      summary: 사용자 역할 변경
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-004
      - REQ-017
      - REQ-021
      - REQ-027
      x-business-rules:
      - POSITION 역할은 직접 변경할 수 없다
      - 승인자와 유효기간은 유효한 계정과 날짜여야 한다
      x-side-effects:
      - user_role의 유효기간과 승인자를 갱신한다
      - change_history에 변경 전후 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 사용자 역할 변경 요청 -> 후속 조회에서 관련 table 변경값이 ApiResponse.data에 표시된다'
      - 'auth: 세션 없는 사용자 역할 변경 요청 -> HTTP 401과 ApiError.message가 반환된다'
      - 'validation: 필수 식별자 누락 -> ApiError.fields에 해당 field 오류가 포함된다'
      - 'business: POSITION 역할은 직접 변경할 수 없다 위반 -> 대상 table의 기존 상태가 변경되지 않는다'
      - 'side-effect: 사용자 역할 변경 성공 -> change_history 또는 관련 상태 row가 생성된다'
      responses:
        '200':
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 검증 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      parameters:
      - name: assignmentId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      x-roles:
      - authenticated
    delete:
      operationId: revokeUserRole
      summary: 수동 역할 회수
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-004
      - REQ-017
      - REQ-027
      x-business-rules:
      - POSITION 역할은 직접 회수할 수 없다
      - 이미 REVOKED인 역할은 중복 회수하지 않는다
      x-side-effects:
      - user_role.status를 REVOKED로 갱신한다
      - change_history에 회수 사유를 기록한다
      x-required-tests:
      - 'happy: 유효한 수동 역할 회수 요청 -> 후속 조회에서 관련 table 변경값이 ApiResponse.data에 표시된다'
      - 'auth: 세션 없는 수동 역할 회수 요청 -> HTTP 401과 ApiError.message가 반환된다'
      - 'validation: 필수 식별자 누락 -> ApiError.fields에 해당 field 오류가 포함된다'
      - 'business: POSITION 역할은 직접 회수할 수 없다 위반 -> 대상 table의 기존 상태가 변경되지 않는다'
      - 'side-effect: 수동 역할 회수 성공 -> change_history 또는 관련 상태 row가 생성된다'
      responses:
        '200':
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 검증 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      x-state-transitions:
      - ACTIVE -> REVOKED
      parameters:
      - name: assignmentId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      x-roles:
      - authenticated
...
  /api/code-groups/{groupId}/detail-codes:
    get:
      operationId: listDetailCodes
      summary: 상세코드 조회
      x-related-requirements:
      - REQ-009
      - REQ-018
      - REQ-027
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 검증 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - sessionCookie: []
      parameters: *id004
      x-roles:
      - authenticated
    post:
      operationId: createDetailCode
      summary: 상세코드 등록
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-009
      - REQ-017
      - REQ-018
      - REQ-021
      - REQ-027
      x-business-rules:
      - code_value는 group_id 안에서 중복될 수 없다
      - effective_end_date는 effective_start_date보다 이전일 수 없다
      x-side-effects:
      - detail_code row를 생성한다
      - change_history에 상세코드 등록 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 상세코드 등록 요청 -> 후속 조회에서 관련 table 변경값이 ApiResponse.data에 표시된다'
      - 'auth: 세션 없는 상세코드 등록 요청 -> HTTP 401과 ApiError.message가 반환된다'
      - 'validation: 필수 식별자 누락 -> ApiError.fields에 해당 field 오류가 포함된다'
      - 'business: code_value는 group_id 안에서 중복될 수 없다 위반 -> 대상 table의 기존 상태가 변경되지 않는다'
      - 'side-effect: 상세코드 등록 성공 -> change_history 또는 관련 상태 row가 생성된다'
      responses:
        '200':
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 검증 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      x-roles:
      - authenticated
  /api/code-groups/{groupId}/detail-codes/{codeValue}:
    put:
      operationId: updateDetailCode
      summary: 상세코드 수정
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-009
      - REQ-017
      - REQ-018
      - REQ-019
      - REQ-021
      - REQ-027
      x-business-rules:
      - 상위코드는 같은 group_id 안의 code_value여야 한다
      - 사용 중인 상세코드는 물리삭제보다 INACTIVE를 우선한다
      x-side-effects:
      - detail_code 정보를 갱신한다
      - change_history에 상세코드 수정 전후 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 상세코드 수정 요청 -> 후속 조회에서 관련 table 변경값이 ApiResponse.data에 표시된다'
      - 'auth: 세션 없는 상세코드 수정 요청 -> HTTP 401과 ApiError.message가 반환된다'
      - 'validation: 필수 식별자 누락 -> ApiError.fields에 해당 field 오류가 포함된다'
      - 'business: 상위코드는 같은 group_id 안의 code_value여야 한다 위반 -> 대상 table의 기존 상태가 변경되지 않는다'
      - 'side-effect: 상세코드 수정 성공 -> change_history 또는 관련 상태 row가 생성된다'
      responses:
        '200':
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 검증 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
      - name: codeValue
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      x-roles:
      - authenticated
  /api/code-groups/{groupId}/detail-codes/tree:
    get:
      operationId: getDetailCodeTree
      summary: 상세코드 계층 조회
      x-related-requirements:
      - REQ-009
      - REQ-027
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 검증 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - sessionCookie: []
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
      - *id005
      - *id006
      - *id007
      x-roles:
      - authenticated
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
  schemas:
    ApiError:
      type: object
      required:
      - message
      - fields
      properties:
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
    ApiResponse:
      type: object
      required:
      - success
      properties:
        success:
          type: boolean
        data: {}
        error:
          $ref: '#/components/schemas/ApiError'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
...